Hack to fix mapping axes that jump from 0 to 100%
This addresses issues with things such as d-pads that
are represented as X/Y axes.  The mapper's logic for
dealing with jittery stick/trigger data was reading
the motion as the first event near the 'resting position'
of the axis, and treating the first press as what
'no press' looks like.